fix: not exposing new session function used in test server
