+++ test/test-server/setup-for-test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cp ./additional-files/new_session_for_test_server.go ../../recipe/session/
